Change : update libav, thanks to Alexins; Fix : crash on x64 and YV12 output; Change : remove some C4018 warning;
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4114 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/Subtitles/TextFile.cpp
@@ -150,7 +150,7 @@ ULONGLONG CTextFile::Seek(LONGLONG lOff, UINT nFrom)
 			break;
 	}
 
-	lOff = max(min(lOff, len), 0) + m_offset;
+	lOff = max(min((ULONGLONG)lOff, len), 0) + m_offset;
 
 	pos = CStdioFile::Seek(lOff, begin) - m_offset;
